+++ pkg/dataplane/maps.go
@@ -11,69 +11,6 @@ import (
 	"github.com/cilium/ebpf"
 )
 
-// SetScreenConfig writes a screen profile configuration entry.
-func (m *Manager) SetScreenConfig(profileID uint32, cfg ScreenConfig) error {
-	zm, ok := m.maps["screen_configs"]
-	if !ok {
-		return fmt.Errorf("screen_configs map not found")
-	}
-	return zm.Update(profileID, cfg, ebpf.UpdateAny)
-}
-
-// ClearScreenConfigs zeroes all screen_configs entries.
-func (m *Manager) ClearScreenConfigs() error {
-	zm, ok := m.maps["screen_configs"]
-	if !ok {
-		return fmt.Errorf("screen_configs map not found")
-	}
-	empty := ScreenConfig{}
-	for i := uint32(0); i < 64; i++ {
-		zm.Update(i, empty, ebpf.UpdateAny)
-	}
-	return nil
-}
-
-// UpdateSessionCountSrc writes a per-source-IP session count entry.
-func (m *Manager) UpdateSessionCountSrc(key SessionCountKey, count uint32) error {
-	zm, ok := m.maps["session_count_src"]
-	if !ok {
-		return fmt.Errorf("session_count_src map not found")
-	}
-	val := SessionCountValue{Count: count}
-	return zm.Update(key, val, ebpf.UpdateAny)
-}
-
-// UpdateSessionCountDst writes a per-destination-IP session count entry.
-func (m *Manager) UpdateSessionCountDst(key SessionCountKey, count uint32) error {
-	zm, ok := m.maps["session_count_dst"]
-	if !ok {
-		return fmt.Errorf("session_count_dst map not found")
-	}
-	val := SessionCountValue{Count: count}
-	return zm.Update(key, val, ebpf.UpdateAny)
-}
-
-// ClearSessionCounts deletes all entries from the session count maps.
-func (m *Manager) ClearSessionCounts() error {
-	for _, name := range []string{"session_count_src", "session_count_dst"} {
-		zm, ok := m.maps[name]
-		if !ok {
-			continue
-		}
-		var key SessionCountKey
-		var val []byte
-		iter := zm.Iterate()
-		var keys []SessionCountKey
-		for iter.Next(&key, &val) {
-			keys = append(keys, key)
-		}
-		for _, k := range keys {
-			zm.Delete(k)
-		}
-	}
-	return nil
-}
-
 // SetMirrorConfig writes a port-mirroring entry for the given ingress ifindex.
 func (m *Manager) SetMirrorConfig(ifindex int, mirrorIfindex int, rate uint32) error {
 	zm, ok := m.maps["mirror_config"]
@@ -150,25 +87,6 @@ func (m *Manager) IncrementGlobalCounter(index uint32, delta uint64) error {
 	m.userspaceCounterOffsets[index] += delta
 	m.mu.Unlock()
 	return nil
-}
-
-// ReadFloodCounters reads the per-CPU flood state for a zone and sums them.
-func (m *Manager) ReadFloodCounters(zoneID uint16) (FloodState, error) {
-	zm, ok := m.maps["flood_counters"]
-	if !ok {
-		return FloodState{}, fmt.Errorf("flood_counters map not found")
-	}
-	var perCPU []FloodState
-	if err := zm.Lookup(uint32(zoneID), &perCPU); err != nil {
-		return FloodState{}, err
-	}
-	var total FloodState
-	for _, fs := range perCPU {
-		total.SynCount += fs.SynCount
-		total.ICMPCount += fs.ICMPCount
-		total.UDPCount += fs.UDPCount
-	}
-	return total, nil
 }
 
 // ReadInterfaceCounters reads the per-CPU interface counter values and sums them.
